feature: specialize functions at certain types
instead of specializing a type being built by passing a specific value
we can know pass a specific function. That function will fetch its inputs as
any other function.
The API for specialization has been made more orthogonal and now
requires a `val` or a `fun` as a parameter (or `valTo/funTo`)